Pre-refactor
Due to the usage of BeautifulSoup's finder to immediately move to an element, lots of try/excepts need to handle cases in which BS throws an AttributeError for trying to move along a NoneType.

Refactor intends to handle this case more elegantly by checking for NoneTyped finder results rather than immediately trying to step forward from the finder result.